Refactor files to have a better structure
